otap: rename AttributeProcessor -> AttributesProcessor; rename config -> Config; align rename fields to source_key/destination_key; register with new URN; remove old module
